Delete useAuth.
